# conf/openapi.yaml
swagger: "2.0"
servers:
  - url: http://localhost:5000/
    description: Local
info:
  description: This is the API for Elearning website
  version: "1.0.0"
  title: E-learning API
  contact:
    email: apiElearning@gmail.com
...
paths:
...
  /courses/get/{courseId}:
    get:
      tags:
        - about courses
      summary: get course by ID
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: search the course by ID 
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/courses'
        '400':
          description: course ID doesn't found
...
components:
  schemas:
...
    courses:
      description: course given by an instructor
      type: object
      required:
        - uuid
        - instructorID
        - title
        - description
        - price
        - requirements
        - video
        - state
        - chapters
        - category
        - createdAt
      properties:
        oid:
          description: UUID of the course 
          type: string
          format: uuid
          example: 54ee07dbc3f29cc751a23829
        instructorID:
          description: UUID of the instructor given the course 
          type: string
          format: uuid
        chapters:
          description: array of chapters
          type: array
          items:
            $ref: '#/components/schemas/chapters'
        createdAt:
          description: data creation of the room
          type: string
          format: date
    chapters:
      description: chapter of a course
      type: object
      required:
        - title
        - description
        - content
      properties:
        title:
          description: title of the chapter
          type: string
          example: inroduction to Python
        description:
          description: desciption of the chapter
          type: string
          example: in this chapter ....
        lessions:
          description: array of lession
          type: array
          items:
            $ref: '#/components/schemas/lessions'
    lessions:
      description: lession of a chapter
      type: object
      required:
        - title
        - file
        - fileType
      properties:
        title:
          description: title of the lession
          type: string
          example: inroduction to Python
        file:
          description: URL of file attached to lession
          type: string
        fileType:
          description: type of the file; can be video of pdf
          type: string
          enum: [video, pdf]